Store new inference metrics calculation
